+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/file/converter/ExcelFileConverter.kt
@@ -1,8 +1,8 @@
-package team.comit.simtong.domain.file.convertor
+package team.comit.simtong.domain.file.converter
 
 import org.springframework.web.multipart.MultipartFile
-import team.comit.simtong.domain.file.convertor.FileExtensions.XLS
-import team.comit.simtong.domain.file.convertor.FileExtensions.XLSX
+import team.comit.simtong.domain.file.converter.FileExtensions.XLS
+import team.comit.simtong.domain.file.converter.FileExtensions.XLSX
 
 /**
  *
