+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/file/converter/FileExtensions.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.file.convertor
+package team.comit.simtong.domain.file.converter
 
 /**
  *
